requirements.txt
flet
pyrebase4
requests
